requirements.txt
fastapi
uvicorn
python-dotenv
langchain-community
langchain-huggingface
langchain-core
faiss-cpu
huggingface-hub
pydantic
requests